+++ tapis-jobsapi/CHANGELOG.md
@@ -14,6 +14,7 @@ https://tapis-project.github.io/live-docs
 ### New Features:
 1. Allow docker --entrypoint container parameter to be specified.
 2. Kill FORK jobs whose maxMinutes have expired.
+3. Resolve macros that appear in archiveFilter lists.
 
 ### Bug fixes:
 1. Fix ZIP bug when staging applications and extraneous text is written to stdout.